+++ client/pages/h5-example/index.js
@@ -7,7 +7,7 @@ import {urlContext} from '../../utils/config';
 import H5ExamplePage from './H5ExamplePage';
 import Detail from './detail/components/Detail';
 
-const routes = () => {
+const container = () => {
   return (
     <H5ExamplePage>
       <Route path="/detail/:projectId" component={Detail}/>
@@ -16,6 +16,6 @@ const routes = () => {
 };
 
 render(
-  <Root routes={routes} reducers={reducers} basename={`${urlContext}/h5-example`}/>,
+  <Root container={container} reducers={reducers} basename={`${urlContext}/h5-example`}/>,
   document.getElementById('layout')
 );