+++ client/pages/home/components/HomeRoute2.js
@@ -27,9 +27,7 @@ export default class HomeRoute2 extends Component {
     } = this.props;
 
     return (
-      <div className="m-t-6">
-        <h5 className="theme-secondary">{home.getIn(['homeRoute', 'content'])}，路由 pathname： {pathname}</h5>
-      </div>
+      <h5 className="theme-secondary">{home.getIn(['homeRoute', 'content'])}，路由 pathname： {pathname}</h5>
     );
   }
 }