+++ client/pages/page-1/reducers.js
@@ -1,5 +1,6 @@
 import {combineReducers} from 'redux-immutable';
 import toast from '../../common/reducers/toast';
+import loading from '../../common/reducers/loading';
 import caches from '../../common/reducers/caches';
 import module1 from './module-1/reducer';
 import module2 from './module-2/reducer';
@@ -7,6 +8,7 @@ import module2 from './module-2/reducer';
 export default combineReducers({
   toast,
   caches,
+  loading,
   module1,
   module2
 });